gallery_app/backend/requirements.txt
# requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Pillow==10.0.1
Werkzeug==2.3.7
